Rename scroll.dart to drag.dart
The gesture is now called "drag". The library name should match.
